requirements.txt
altgraph==0.17
Babel==2.9.0
bleach==3.3.0
blinker==1.4
cffi==1.14.5
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Admin==1.5.7
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security-Too==4.0.0
Flask-WTF==0.14.3
future==0.18.2
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
loginform==1.2.0
lxml==4.6.2
MarkupSafe==1.1.1
packaging==20.9
passlib==1.7.4
pefile==2019.4.18
pip-review==1.1.0
pycparser==2.20
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pyparsing==2.4.7
pytz==2021.1
pywin32-ctypes==0.2.0
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.23
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3
